use native types name
eliminating _t typedef use

+++ set.h
@@ -13,8 +13,8 @@ namespace ici
  */
 struct set : object
 {
-    int      s_nels;            /* How many slots used. */
-    int      s_nslots;          /* How many slots allocated. */
+    size_t  s_nels;            /* How many slots used. */
+    size_t  s_nslots;          /* How many slots allocated. */
     object **s_slots;
 };
